+++ Lab4/src/localization/LCDInfo.java
@@ -29,9 +29,9 @@ public class LCDInfo implements TimerListener{
 		LCD.clear();
 		LCD.drawString("X: ", 0, 0);
 		LCD.drawString("Y: ", 0, 1);
-		LCD.drawString("H: ", 0, 2);
-		LCD.drawInt((int)(pos[0] * 10), 3, 0);
-		LCD.drawInt((int)(pos[1] * 10), 3, 1);
-		LCD.drawInt((int)pos[2], 3, 2);
+		LCD.drawString("theta: ", 0, 2);
+		LCD.drawInt((int)(pos[0]), 3, 0);
+		LCD.drawInt((int)(pos[1]), 3, 1);
+		LCD.drawInt((int)pos[2], 8, 2);
 	}
 }
